+++ src/codec.js
@@ -47,7 +47,8 @@ const enc = new Map([
 
   [opcodes.OpenDoor, encoder.OpenDoor],
   [opcodes.SetPCControl, encoder.SetPCControl],
-  [opcodes.SetInterlock, encoder.SetInterlock]
+  [opcodes.SetInterlock, encoder.SetInterlock],
+  [opcodes.ActivateKeypads, encoder.ActivateKeypads]
 ])
 
 /**
@@ -75,6 +76,7 @@ const dec = new Map([
   [0x98, decoder.GetTimeProfile],
   [0xa0, decoder.SetPCControl],
   [0xa2, decoder.SetInterlock],
+  [0xa4, decoder.ActivateKeypads],
   [0xa6, decoder.ClearTaskList],
   [0xa8, decoder.AddTask],
   [0xac, decoder.RefreshTaskList],